% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.sprk.R
\name{hist.sprk}
\alias{hist.sprk}
\title{A Histogram Function}
\usage{
\method{hist}{sprk}(df, col)
}
\arguments{
\item{df}{The name of the DataFrame.}

\item{col}{The name of the column, character string.}
}
\description{
This function generates a frequency histogram
of single columns from a SparkR DataFrame.
}
\examples{
hist.sprk(iris, "Sepal_Length")
}
\keyword{data-exploration,}
\keyword{frequency,}
\keyword{table,}
\keyword{visualisation}

