% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.sR.R
\name{hist.sR}
\alias{hist.sR}
\title{A Histogram Function}
\usage{
\method{hist}{sR}(df, col)
}
\arguments{
\item{df}{The name of the DataFrame.}

\item{col}{The name of the column, character string.}
}
\description{
This function generates a frequency histogram
of a single column from a SparkR DataFrame.
}
\examples{
hist.sR(iris, "Sepal_Length")
}
\keyword{data-exploration,}
\keyword{frequency,}
\keyword{sparkR}
\keyword{table,}
\keyword{visualisation,}

