% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.sR.R
\name{median.sR}
\alias{median.sR}
\title{A Median Function}
\usage{
\method{median}{sR}(df, col)
}
\arguments{
\item{df}{The name of the DataFrame.}

\item{col}{The name of the column, character string.}
}
\description{
This function computes the sample median
of a single column from a SparkR DataFrame.
}
\examples{
median.sR(iris, "Sepal_Length")
}
\keyword{data-exploration,}
\keyword{frequency,}
\keyword{median,}
\keyword{sparkR,}
\keyword{statistics}
\keyword{summary}

