% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.sR.R
\name{mode.sR}
\alias{mode.sR}
\title{A Mode Function}
\usage{
mode.sR(df, col)
}
\arguments{
\item{df}{The name of the DataFrame.}

\item{col}{The name of the column, character string.}
}
\description{
This function computes the sample mode
of a single column from a SparkR DataFrame.
}
\examples{
mode.sR(iris, "Sepal_Length")
}
\keyword{data-exploration,}
\keyword{frequency,}
\keyword{mode,}
\keyword{sparkR,}
\keyword{statistics}
\keyword{summary}

