% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{A Tabulate Function}
\usage{
tab(df, col)
}
\arguments{
\item{df}{The name of the DataFrame.}

\item{col}{The name of the column, character string.}
}
\description{
This function allows you attain frequency tables
of single columns from a SparkR DataFrame.
}
\examples{
tab(iris, "Sepal_Length")
}
\keyword{data-exploration,}
\keyword{frequency}
\keyword{table,}

