% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.sR.R
\name{table.sR}
\alias{table.sR}
\title{A Tabulate Function}
\usage{
table.sR(df, col, col2 = col)
}
\arguments{
\item{df}{The name of the DataFrame.}

\item{col}{The name of the column, character string.}
}
\description{
This function allows you to attain a frequency table
of a single or two columns from a SparkR DataFrame.
The output is meant to replicate that of base::table().
}
\examples{
table.sR(iris, "Sepal_Length")
table.sR(iris, "Sepal_Length", "Sepal_Width")
}
\keyword{data-exploration,}
\keyword{frequency,}
\keyword{sparkR}
\keyword{table,}

